Choose kernel based on integer instruction support, disable vectorization in kernel

+++ boot/elf64.cc
@@ -262,7 +262,7 @@ bool elf64_run(char const *filename)
             vaddr += chunk_size;
 
             progress_bar_draw(20, 10, 70,
-                              100 * (double)done_bytes / (double)total_bytes);
+                              100 * done_bytes / total_bytes);
         }
 
         // Clear modified bits if uninitialized data
